updated bash script

+++ steamfix.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
-
-echo "Fixing Steam...."
-
-find ~/.local/share/Steam/ \( -name "libgcc_s.so*" -o -name "libstdc++.so*" -o -name "libxcb.so*" -o -name "libgpg-error.so*" \) -print -delete
-
-echo "Done!!!"
+sleep 1
+if [ ! -d "~/.local/share/Steam" ];
+then
+	echo "No Steam Folder Present"
+else
+	echo "Fixing Steam..."
+	sleep 1	
+	find ~/.local/share/Steam/ \( -name "libgcc_s.so*" -o -name "libstdc++.so*" -o -name "libxcb.so*" -o -name "libgpg-error.so*" \) -print -delete
+	echo "Done!!!"
+fi
+sleep 1
